Update Pong.cpp
Modified the basic ball speed value from 0.08 to 0.10. Values between 0.09 to 0.12 should be considered for the initial value. 0.08 worked well before the change to using angles, but it no longer does because the change to use angles resulted in the actual velocity being lower. So, increasing it up between 0.09 and 0.12 gets it closer to the original ball speed.